* Removed Json.NET implementation * Added SQLite Database * Removed all instances of Json.Net

+++ CryptoLedger/Asset.cs
@@ -1,14 +1,16 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using System.Text.Json;
+using System.Text.Json.Serialization;
 
 namespace CryptoLedger
 {
     public class Asset
     {
         public string Ticker { get; set; }
-        public float Amount { get; set; }
-        public int Invested { get; set; }
+        public decimal Amount { get; set; }
+        public decimal Invested { get; set; }
         public string Wallet { get; set; }
         public bool isStaked { get; set; }
     }
